+++ components/shipstation/src/inputs/shipments/listShipments.ts
@@ -1,22 +1,8 @@
-import { connectionInput, page, pageSize } from "../common";
-import {
-  createDateEnd,
-  createDateStart,
-  recipientCountryCode,
-  recipientName,
-  shipDateEnd,
-  shipDateStart,
-  trackingNumber,
-} from "./common";
+import { connectionInput, pagination } from "../common";
+import { dateRangeFilters, shipmentFilters } from "./common";
 export const listShipmentsInputs = {
   connectionInput,
-  trackingNumber,
-  createDateStart,
-  createDateEnd,
-  shipDateStart,
-  shipDateEnd,
-  recipientName,
-  recipientCountryCode,
-  page,
-  pageSize,
+  shipmentFilters,
+  dateRangeFilters,
+  pagination,
 };